微擎新版：微官网幻灯片、快捷菜单

+++ web/source/site/slide.ctrl.php
@@ -7,7 +7,7 @@ defined('IN_IA') or exit('Access Denied');
 $do = !empty($do) ? $do : 'display';
 $do = in_array($do, array('display', 'post', 'delete')) ? $do : 'display';
 
-if ($do == 'display') {
+if ($do == 'display' && $_W['isajax']) {
 	$pindex = max(1, intval($_GPC['page']));
 	$psize = 10;
 	$condition = '';
@@ -25,27 +25,25 @@ if ($do == 'display') {
 	message($list, 'ajax', 'success');
 }
 
-if ($do == 'post') {
-	if ($_W['isajax'] && $_W['ispost']) {
-		$post = $_GPC['__input'];
-		foreach ($post['slide'] as $key => $val) {
-			if(empty($val['thumb'])){
-				message('幻灯图片不可为空', 'ajax', 'error');
-			}
+if ($do == 'post' && $_W['isajax'] && $_W['ispost']) {
+	$post = $_GPC['__input'];
+	foreach ($post['slide'] as $key => $val) {
+		if(empty($val['thumb'])){
+			message('幻灯图片不可为空', 'ajax', 'error');
 		}
-		$multiid = intval($post['multiid']);
-		pdo_fetch("DELETE FROM ".tablename('site_slide')." WHERE uniacid = :uniacid AND multiid = :multiid" , array(':uniacid' => $_W['uniacid'],':multiid' => $multiid));
-		foreach ($post['slide'] as  $value) {
-			$data = array(
-				'uniacid' => $_W['uniacid'],
-				'multiid' => $multiid,
-				'title' => $value['title'],
-				'url' => $value['url'],
-				'thumb' => $value['thumb'],
-				'displayorder' => intval($value['displayorder']),
-			);
-			pdo_insert('site_slide', $data);
-		}
-		message('幻灯片保存成功！', 'ajax', 'success');
 	}
+	$multiid = intval($post['multiid']);
+	pdo_fetch("DELETE FROM ".tablename('site_slide')." WHERE uniacid = :uniacid AND multiid = :multiid" , array(':uniacid' => $_W['uniacid'],':multiid' => $multiid));
+	foreach ($post['slide'] as  $value) {
+		$data = array(
+			'uniacid' => $_W['uniacid'],
+			'multiid' => $multiid,
+			'title' => $value['title'],
+			'url' => $value['url'],
+			'thumb' => $value['thumb'],
+			'displayorder' => intval($value['displayorder']),
+		);
+		pdo_insert('site_slide', $data);
+	}
+	message('幻灯片保存成功！', 'ajax', 'success');
 }
